refactor(button): refactor naming for styles and files.
