+++ src/components/Button/styles.ts
@@ -1,6 +1,6 @@
 import { css } from '@emotion/css';
 import ColorSystem from '../ColorSystem/index';
-import { generateDefaultPrimarySecondaryButtonStyles } from './utils/StylesGenerator';
+import { getDefaultButtonStyles } from './utils/stylesGenerator';
 
 const SMALL_BTN = '0.25rem 0.5rem';
 const MEDIUM_BTN = '0.5rem 1rem';
@@ -24,65 +24,69 @@ const mainStyles = {
 };
 
 export const styButton = {
-    Primary: css({
-        ...mainStyles,
-        ...generateDefaultPrimarySecondaryButtonStyles(
-            `solid 2px ${ColorSystem.BLUE500}`,
-            ColorSystem.BLUE500,
-            ColorSystem.GREY100
-        ),
-        ':hover': {
-            backgroundColor: ColorSystem.BLUE400,
-            border: `solid 2px ${ColorSystem.BLUE400}`,
-            cursor: 'pointer',
-        },
-        '&.disabled': {
-            backgroundColor: ColorSystem.GREY300,
-            border: `solid 2px ${ColorSystem.GREY300}`,
-            color: ColorSystem.GREY100,
-            cursor: 'not-allowed',
-        },
-    }),
+    Primary: (customColor = ColorSystem.BLUE500) =>
+        css({
+            ...mainStyles,
+            ...getDefaultButtonStyles(
+                `solid 2px ${customColor}`,
+                customColor,
+                ColorSystem.GREY100
+            ),
+            ':hover': {
+                backgroundColor: ColorSystem.BLUE400,
+                border: `solid 2px ${ColorSystem.BLUE400}`,
+                cursor: 'pointer',
+            },
+            '&.disabled': {
+                backgroundColor: ColorSystem.GREY300,
+                border: `solid 2px ${ColorSystem.GREY300}`,
+                color: ColorSystem.GREY100,
+                cursor: 'not-allowed',
+            },
+        }),
 
-    Secondary: css({
-        ...mainStyles,
-        ...generateDefaultPrimarySecondaryButtonStyles(
-            `solid 2px ${ColorSystem.BLUE500}`,
-            ColorSystem.GREY100,
-            ColorSystem.BLUE500
-        ),
-        '&:hover': {
-            backgroundColor: ColorSystem.BLUE400,
-            border: `solid 2px ${ColorSystem.BLUE400}`,
-            cursor: 'pointer',
-            color: ColorSystem.GREY100,
-        },
-        '&.disabled': {
-            border: `solid 2px ${ColorSystem.GREY300}`,
-            color: ColorSystem.GREY300,
-            cursor: 'not-allowed',
+    Secondary: (customColor = ColorSystem.BLUE500) =>
+        css({
+            ...mainStyles,
+            ...getDefaultButtonStyles(
+                `solid 2px ${customColor}`,
+                ColorSystem.GREY100,
+                customColor
+            ),
             '&:hover': {
-                backgroundColor: ColorSystem.GREY100,
+                backgroundColor: ColorSystem.BLUE400,
+                border: `solid 2px ${ColorSystem.BLUE400}`,
+                cursor: 'pointer',
+                color: ColorSystem.GREY100,
+            },
+            '&.disabled': {
+                border: `solid 2px ${ColorSystem.GREY300}`,
                 color: ColorSystem.GREY300,
                 cursor: 'not-allowed',
+                '&:hover': {
+                    backgroundColor: ColorSystem.GREY100,
+                    color: ColorSystem.GREY300,
+                    cursor: 'not-allowed',
+                },
             },
-        },
-    }),
+        }),
 
-    Link: css({
-        border: 'none !important',
-        backgroundColor: 'transparent !important',
-        ':hover': {
-            cursor: 'pointer',
-        },
-        '&:hover': {
-            color: ColorSystem.BLUE400,
-        },
-        '&.disabled': {
-            color: ColorSystem.GREY300,
-            '&:hover': {
-                cursor: 'not-allowed',
+    Link: (customColor = ColorSystem.BLUE400) =>
+        css({
+            border: 'none',
+            backgroundColor: 'transparent',
+            color: customColor,
+            ':hover': {
+                cursor: 'pointer',
             },
-        },
-    }),
+            '&:hover': {
+                color: ColorSystem.GREY300,
+            },
+            '&.disabled': {
+                color: customColor,
+                '&:hover': {
+                    cursor: 'not-allowed',
+                },
+            },
+        }),
 };
